pyproject.toml
[tool.poetry]
name = "rul-adapt"
version = "0.0.0"
description = "A collection of unsupervised domain adaption approaches for RUL estimation."
authors = ["Krokotsch, Tilman <tilman.krokotsch@tu-berlin.de>"]
license = "MIT"
readme = "README.md"
homepage = "https://krokotsch.eu/rul-adapt"
repository = "https://github.com/tilman151/rul-adapt"
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering "
]
packages = [{include = "rul_adapt"}]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.8.0.post1"
rul-datasets = "^0.10.4"
tqdm = "^4.62.2"
hydra-core = "^1.3.1"
pywavelets = "^1.4.1"
dtaidistance = "^2.3.10"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^22.8.0"}
flake8 = "^5.0.4"
mypy = "^1.0.0"
pytest = "^7.1.3"
tensorboard = "^2.11.0"
pytest-mock = "^3.10.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.1"
mkdocs-material = "^8.5.7"
mkdocstrings = {extras = ["python"], version = "^0.19.0"}
mkdocs-gen-files = "^0.4.0"
mkdocs-literate-nav = "^0.5.0"
mkdocs-section-index = "^0.3.4"
jupyter = "^1.0.0"
matplotlib = "^3.7.0"

[tool.pytest.ini_options]
markers = ["integration: does a full train-test run"]

[tool.mypy]
plugins = "numpy.typing.mypy_plugin"
exclude = ".venv"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry_bumpversion.file."rul_datasets/__init__.py"]